A3 phaseobserver (#53)
* Added abstract classes (interfaces) for Game Observers

* Implemented Observable class and updated Player.cpp to be observable.

* Implemented Game stats observer

* Initial Update

Due to some weird obscure error I wasn't able to debug, the code structure was rewritten and simplified. This is a very draft version of strategy pattern implementation

* Tweaked the observer and observable classes. Added PhaseObserver class.

* Implemented all necessary changes to classes other than Strategy

Automated strategy is not yet complete. But changes to code bases are done (I sincerely hope so)

* Finished implementing Strategy Pattern

* Finished Phase Observer implementation and integration. Included phase observer in player strategy test driver.

* Finalized phase observer. Moved instantiation to GameBuilder.

* Bug fixes in GameBuilder

* A3 strategy pattern (#52)

* Added abstract classes (interfaces) for Game Observers

* Implemented Game stats observer

* Initial Update

Due to some weird obscure error I wasn't able to debug, the code structure was rewritten and simplified. This is a very draft version of strategy pattern implementation

* Implemented all necessary changes to classes other than Strategy

Automated strategy is not yet complete. But changes to code bases are done (I sincerely hope so)

* Finished implementing Strategy Pattern

Co-authored-by: Josh Pham <joshpham.concordia@gmail.com>

* .gitignore fix (#46)

* Implemented Observable class and updated Player.cpp to be observable.

* Implemented Game stats observer

* Initial Update

Due to some weird obscure error I wasn't able to debug, the code structure was rewritten and simplified. This is a very draft version of strategy pattern implementation

* Tweaked the observer and observable classes. Added PhaseObserver class.

* Implemented all necessary changes to classes other than Strategy

Automated strategy is not yet complete. But changes to code bases are done (I sincerely hope so)

* Finished implementing Strategy Pattern

* Finished Phase Observer implementation and integration. Included phase observer in player strategy test driver.

* Finalized phase observer. Moved instantiation to GameBuilder.

* Bug fixes in GameBuilder

* Fixed small rebase error in Player.h

Co-authored-by: Josh Pham <joshpham.concordia@gmail.com>
Co-authored-by: goldenfire-yue <jamesyuechao@gmail.com>
Co-authored-by: goldenfire-yue <35940665+goldenfire-yue@users.noreply.github.com>
Co-authored-by: jonZlotnik <me@jonzlotnik.info>